Time: 6 ms (86.24%), Space: 17.3 MB (72.73%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -22,13 +22,13 @@ public:
             int first,last;
             int mini=qq.front().second;
             for(int i=0;i<size;i++){
-                long long currindex=qq.front().second-mini;
+                int currindex=qq.front().second-mini;
                 TreeNode* curr=qq.front().first;
                 qq.pop();
                 if(i==0)first=currindex;
                 if(i==size-1)last=currindex;
-                if(curr->left!=NULL){qq.push({curr->left,currindex*2+1});}
-                if(curr->right!=NULL){qq.push({curr->right,currindex*2+2});}
+                if(curr->left!=NULL){qq.push({curr->left,(long long)currindex*2+1});}
+                if(curr->right!=NULL){qq.push({curr->right,(long long)currindex*2+2});}
             }
             ans=max(ans,last-first+1);
         }
